# Optimized matrix-to-raster for t/tensor? inputs, GRID-340