... yielding +50% throughput improvement.

## Purpose
Faster Geotiff writing in GridFire.

## Related Issues
Closes GRID-340.

## Submission Checklist
- [x] Commits (_**kinda**_) include the JIRA issue and the `#review` hashtag (e.g. `MAG-### #review <comment>`)
- [x] Code passes linter rules (`clj-kondo --lint src`)
- [x] Tests pass (`clj -M:test`)

## Performance measurements

Measuring the performance of `gridfire.simulations/process-output-layers!` in a configuration that writes only GeoTIFFs.

## Before optimizing

```
       ;id: :gridfire.simulations/write-outputs
       ;pId                         nCalls        Min      50% ≤      90% ≤      95% ≤      99% ≤        Max       Mean   MAD      Clock  Total
       ;
       ;:process-output-layers!         30     3.36s      3.48s      3.93s      4.52s      4.91s      4.91s      3.59s    ±7%     1.79m    100%
       ;
       ;Accounted                                                                                                                 1.79m    100%
       ;Clock                                                                                                                     1.79m    100%
```

![process-output-layers_before-optimizing](https://user-images.githubusercontent.com/5859120/181034690-934fd17b-1801-45fc-b0e4-276d628d8d77.svg)


## After optimizing

```
       ;id: :gridfire.simulations/write-outputs
       ;pId                         nCalls        Min      50% ≤      90% ≤      95% ≤      99% ≤        Max       Mean   MAD      Clock  Total
       ;
       ;:process-output-layers!         30     2.23s      2.31s      2.44s      3.03s      3.78s      3.78s      2.40s    ±6%     1.20m    100%
       ;
       ;Accounted                                                                                                                 1.20m    100%
       ;Clock                                                                                                                     1.20m    100%
```

![process-output-layers_after-optimizing](https://user-images.githubusercontent.com/5859120/181034885-e17e30b8-82a9-4f4c-8f42-78804dd019c6.svg)